@@ -1,6 +1,6 @@
-module github.com/bmaupin/go-epub
+module github.com/boomhut/go-epub
 
-go 1.20
+go 1.25
 
 require (
 	github.com/gabriel-vasile/mimetype v1.4.2
